Support `Command` instructing `Ui`.
Currently there is no way for `Command` to instruct `Ui` to show help,
exit (and in the upcoming use case, switch to `ClientView`). The former
two are currently done by passing data values (in boolean) to
`MainWindow` logic to handle. This is not a particularly clean way.
Furthermore, switching to `ClientView` (and eventually dealing with
todo, meetings and other front-end) requires a better way to instruct
Ui.

This commit will remove the earlier ways of instructing show help and
exit, and instead rely on a `Consumer<Ui>` field in `CommandResult` to
carry out instructions on `Ui`. This can be extended to handle current
and future `Ui` actions.