Fix failing test cases (Re: `CommandResult`)
Due to the new implementation of instructing `Ui` (no more boolean
values for `showHelp` and `exit`), it is no longer possible for
`CommandResult` to be compared on anything other than the value of its
String `userFeedback`. This commit removes all testing functionality
that needs to compare how `CommandResult` instructs `Ui`.