Create `AttributeTable`.
We want the user to be able to see a table for viewing a `Person`'s
`Collection<>` of `Policy`,`Asset`, `Liability`.

This serves as a generic front end template. In this single component,
we can create all manner of tables that take in a given table config
and a collection of the attribute.

There are key things to note that have been documented partially.